requirements.txt
hexdump
pyelftools
